feat(rime_levers_api): stdbool flavor
