+++ test/flavored_api_test.cc
@@ -21,3 +21,22 @@ TEST(RimeApiStdboolTest, GetContext) {
   ASSERT_TRUE(rime->get_status(test_session, &status));
   ASSERT_FALSE(status.is_composing);
 }
+
+TEST(RimeLeversApiStdboolTest, CustomSettings) {
+  RIME_FLAVORED(RimeApi)* rime = RIME_FLAVORED(rime_get_api)();
+  RimeModule* module = rime->find_module("levers_stdbool");
+  ASSERT_TRUE(bool(module));
+  RIME_FLAVORED(RimeLeversApi)* levers =
+      (RIME_FLAVORED(RimeLeversApi)*)module->get_api();
+  ASSERT_TRUE(bool(levers));
+
+  ASSERT_TRUE(RIME_API_AVAILABLE(levers, custom_settings_init));
+  RimeCustomSettings* custom_settings =
+      levers->custom_settings_init("flavored_api_test", "rime_test");
+  ASSERT_TRUE(bool(custom_settings));
+
+  ASSERT_TRUE(RIME_API_AVAILABLE(levers, customize_bool));
+  ASSERT_TRUE(levers->customize_bool(custom_settings, "test_key", true));
+
+  levers->custom_settings_destroy(custom_settings);
+}
